move AbstractUtils to abstracts package and implement test for issue #626
