+++ core/src/main/scala/org/dbpedia/extraction/nif/WikipediaNifExtractor.scala
@@ -4,7 +4,8 @@ import org.dbpedia.extraction.config.Config
 import org.dbpedia.extraction.config.provenance.DBpediaDatasets
 import org.dbpedia.extraction.ontology.{Ontology, OntologyProperty, RdfNamespace}
 import org.dbpedia.extraction.transform.{Quad, QuadBuilder}
-import org.dbpedia.extraction.util.{AbstractUtils, Language, RecordEntry, RecordSeverity}
+import org.dbpedia.extraction.util.abstracts.AbstractUtils
+import org.dbpedia.extraction.util.{Language, RecordEntry, RecordSeverity}
 import org.dbpedia.extraction.wikiparser.{Namespace, WikiPage}
 import org.dbpedia.extraction.wikiparser.impl.wikipedia.Namespaces
 import org.jsoup.nodes.{Document, Element, Node}
