+++ core/src/test/scala/org/dbpedia/extraction/util/AbstractUtilsTest.scala
@@ -1,4 +1,5 @@
 package org.dbpedia.extraction.util
+import org.dbpedia.extraction.util.abstracts.AbstractUtils
 import org.scalatest.FunSuite
 
 class AbstractUtilsTest extends FunSuite {
