+++ dump/src/test/resources/cv-test-groups.csv
@@ -13,3 +13,4 @@ TEST_NAME,ALL,PRODUCTIVE
 #itsrdf,yes,yes
 #wikidata_IRIs,yes,yes
 #abstracts,yes,yes
+#forward_slash_in_resource_names,yes,yes
